Extract Erlang SDK setup logic from Type into ElixirInternalErlangSdkSetup
 Move configureInternalErlangSdk, promptForMiseErlangSdk, and registerErlangSdk out of the Type companion object into a dedicated ElixirInternalErlangSdkSetup object. This isolates the wizard-driven Erlang linking flow and makes it easier to reason about and test.

 Behavioural changes bundled with the extraction:
 - Replace showChooseDialog with showEditableChooseDialog + InputValidator so cancelling the Erlang SDK picker no longer accidentally selects a valid entry
 - Remove the "Please save settings and reopen this dialog" info message and its associated SKIP_ERLANG_SETTINGS_HINT_KEY - no longer needed

 Update TypeErlangAutoLinkTest call sites accordingly.
